requirements.txt
discord.py==2.3.1
aiohttp==3.8.3
chess==1.9.3
chess.com==2.0.3
cairosvg==2.5.2
DateTime==4.4